Fix mojom enums check in Feature Policy presubmit
The presubmit check reports an error when the feature policy enum is
modified: "*.json5 and *.mojom are out of sync"

This presubmit error is reported even when the files have identical
enum definitions due to a mistake in the error condition. Instead of
comparing the set of json5_enums with the mojom_source_path, compare
with the set of mojom_enums.

BUG=1105600

Change-Id: Ibc80ac41a37a2102e37edbbdbd6a8a77633051fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298225
Reviewed-by: Charlie Hu <chenleihu@google.com>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Charlie Hu <chenleihu@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#788649}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5d7521970b891d375f8a2483c80cf14d292c3a9