Replace Axis with VariationPosition.
SkFontArguments started out with an 'Axis' type to specify the
coordinates arguments to a variable font. However, the individual
positions along an axis are referred to as 'coordinates', and the set of
coordinates is commonly referred to as a 'variation'. The term 'axis' is
used in SkFontParameters to refer to the axes themselves.

Use the newer 'VariationPosition' type so that Skia can remove the
somewhat ambiguously named 'Axis' type.

Change-Id: Iea072cfc9c2bc06efab3b18533db4da17289b310
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298345
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Ben Wagner <bungeman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788672}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74131db2065cfee8e04584194975a6d8ccf11b5c