Convert content::ScreenInfo to a blink::ScreenInfo struct.
The content ScreenInfo struct isn't necessary if we expose the
WebScreenInfo struct in common. This may be eventually moved to a single
sole mojom definition but at this time the struct is used on IPC and
the public blink API so it needs to be in common.

BUG=1097816

Change-Id: I7ee5fb9a4c9e3210b1884d256a9386604174be56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298182
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f08f0b56ab67f6a31b1400b1e13c73f7a82e4920