Devtools: add webtests for new Multi Grid highlight tool
This CL is a follow up to it's parent: https://crrev/c/2278408


This adds webtests to validate the highlightJSON generation for the new
grid tool. However, it is dependent on a change to the elements_test_runner
in the devtools-frontend, and cannot be added until it's parent
lands, the devtools-frontend updates dependencies and the
testrunner file, and those changes are merged back into src.

Bug: 1047356

Change-Id: I483e5f81635b0d0ff3cf8019c2fe98155b5cf960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288337
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789113}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f6a414d10bcf304130bb1f56c3ae25f3aa96e4b