TrialTokenValidator should always null-check the OriginTrialPolicy*
If PolicyGetter() is not initialized, it can be nullptr. This was regressed in
https://chromium-review.googlesource.com/c/chromium/src/+/2293475, which
removed a special-case in OriginTrialContext which declined to initialize
the TrialTokenValidator if there was no PolicyGetter(). Now, we always
create the TrialTokenValidator, but it will early-exit with status
kNotSupported if the PolicyGetter() isn't available.

Bug: 1106240
Change-Id: I4a1188e3866ba225b380c761bd444225c744b719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303560
Auto-Submit: Nate Chapin <japhet@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789349}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50b5fa35d4bcd23e2887317439b36f5b3cfe7003