Prevent crash when page_style_sheet is missing an owner document
According to the crash report, the owner document could be a nullptr.
This CL adds a check if an owner document is available and sets the
execution_context to nullptr if it's not available.
CSSStyleDeclaration::setCSSText that receives the execution context
can handle nullptr values.

Fixed: 1090899
Change-Id: I2e0b1a56e6b25e98c043f576dbd230f50c062516
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300382
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789404}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f57bbec2f4b75e9a49484666ccaca2f7d608564