Set a frame's AXTreeID in the renderer when serializing its AXTreeData
Instead of setting this value in the browser process when an IPC
message from the renderer arrives, assign the AXTreeID of a frame
in the renderer process already when serializing its tree-related
data in an AXTreeData struct.

In order to do this, we add a new WebLocalFrame::GetAXTreeID()
method that will return the AXTreeID of a local frame (as backed
by its associated embedding token), that will be useful in the
future to allow certain use cases to communicate directly with
the renderer, without having to go through the browser process.

Bug: 1094150
Change-Id: Ib2c23f811741f4328184099bb6b9027af9c7e2cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292391
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8139c6f8289910b9a96108b2e22ffe28cbdc2a6