Fix scrollbar scrolling under pinch zoom
Scrollbar scrolling in general had a number of problems when pinch-zoom
(a.k.a. page scale factor) is set. On the compositor side, input points
need to be transformed to the scrollbar relative coordinate space in
order to appropriately calculate deltas, since the calculations are all
done in layer relative coordinates. Additionally the scroll deltas
themselves must be scaled by the page scale factor, since the code that
handles GSUs expects this to be encoded in the deltas sent to
LayerTreeHostImpl. Clamping logic also needs to account for the
distribution of deltas across the inner (visual) viewport and the outer
(layout) viewport.

On the main thread, the injected scroll deltas suffer from the same
issue as the compositor thread - they need to apply the input scale as
well. I've also added the ability to hit test the root viewport
scrollbars, since the hit testing code otherwise will hit elements
that are visually under the scrollbar (the positioning of the scrollbar
is done specially via the OverscrollElasticityTransformNode).
Code in Scrollbar also needs to become aware that it should use the
widget-relative coordinates instead of root frame-relative coordinates
(the latter have been transformed based on scale and offset of the
visual viewport), when operating on the root frame layout viewport.

Bug: 531603
Change-Id: I6eb2ecd135a88228d3f4a11bed29ebb9d788b4fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290966
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#789647}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 25da0a8d274a8ca255715e0679e3e40edf6a4264