DevTools: protocol and backend support for line names in overlay
This introduces a new option (showLineNames) to the
gridHighlightConfig option, passed to overlays.
The option defaults to false. When set to true, it tells the
backend implementation of the overlay domain to
also generate offset information for all of the named lines.

The corresponding frontend CL is
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2292205

Because the frontend CL also adds a new setting that needs
telemetry support, the corresponding enum value is added here.

Bug: 1100385
Change-Id: I0487ae7c5d6aad21c8d6b16bcb3be46a08e0ee07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2291435
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#789773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56ea8daf81b55586a6ffcb8162d87b85062eb83c