Fix ServiceWorker.FetchEvent.QueuingTime metric (CheckOfflineCapability)
ServiceWorker.FetchEvent.QueuingTime is used for the Finch experiment
(CheckOfflineCapability), but it always records a minimum value (0.001)
because the order of a subtraction was wrong. The result of a
subtraction became a minus value and it was rounded to a minimum value.

This CL fixes the order in a subtraction.

Bug: 965802, 1106693
Change-Id: I93b60f301a16caef355e221712372c8e5bcde56f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303236
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789810}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 832f62bc3f414dd6be2442c6f581aa27464216d5