[Push] Pass expirationTime attribute from PushMessagingService to PushManager
This CL adds the capability to pass an expirationTime attribute from
PushMessagingServiceImpl down to PushManager, when PushManager::get
Subscription is called. It is an optional parameter specified in
milliseconds from 1st January 1970 UTC according to the standard.

https://w3c.github.io/push-api/#dfn-subscription-expiration-time

When PushManager::getSubscription is called, expiration_time should be
retrieved in PushMessagingServiceImpl from the app_identifier and sent
back as a callback to PushMessagingManager, which sents it to PushPro-
vider, converts it from a mojom::blink::PushSubscriptionPtr to a
PushSubscription object and returned to PushManager.

The expiration time will be stored in PushMessagingAppIdentifier, which
is currently not implemented yet but WIP. It will then call
PushMessagingServiceImpl::SubscribeEnd with a non-null value for expir-
ation_time and pass the expiration time in SubscriptionInfoCallback.

WIP Issue: 2302610

Change-Id: I1e6c69a930def902cd6f08813de5856ae08729fc
Bug: 1104215
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302192
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Commit-Queue: Viviane Yang <viviy@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#789889}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4b37cb6ba6091d2e52184c7a24a71064252441e