Bidi: RTL that becomes LTR with brackets should be recalculated
For certain cases where a div with RTL direction is set to be LTR, the
algorithm for determining if bidi runs should be recalculated and
reshaped was not properly ran. This is a regression introduced by
https://chromium.googlesource.com/chromium/src/+/25d80f4bead925c2d9071b59a4f32146432d3e9e

The solution is merging the two conditions introduced in the
original patch (NeedsReshape and NeedsCollectInlines) in one
that has been named NeedsReshape.

Bug: 1009654, 1104827
Change-Id: I32fceb75f83b9cf2aac3ef982675a96a3ef4d3c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297432
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#789940}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 14045d11e661ca4793e29763b0826743983f2dee