inspector: Add extra trace arguments to console events
We have a lab metric to track console error messages in telemetry
stories, and when the count changes it is quite difficult to figure
out what the new message is. This CL adds the error message and the url
to the trace event. We will surface this information in the UI for
easier debugging.

Bug: 1086962
Change-Id: Ibeb739c7624941f7437f39e5087fbe574a782cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302837
Commit-Queue: Deep Roy <dproy@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7e438c810ad493cc32752fa72785e819a1b65a8