Remove invalidation code when creating new PaintLayers.
This code was originally added in [1]. The web_test added in that
CL still passes, as does a subsequent related unittest added in another
CL (PaintLayerPainterTest, PaintPhasesUpdateOnLayerAddition).

I'm not sure what caused these to keep passing with the code removed -
there is a reference in [1] to re-use of DisplayItemClient pointers,
perhaps that was improved since then?

[1] https://codereview.chromium.org/1403963002

The bug reported in issue 541544 also does not reproduce with this CL
applied.

Bug: 1104794

Change-Id: I91addcea57c217dd1f971fcc4dbae0cd518c7011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307610
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790106}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4a8929e1d1165ce2c8b81c7d991f93827e41752