[Code health] Fix a float-cast-overflow in WheelEvent
The problem is in the constructor of this class, which calls
NegateIfPossible that takes a int32_t. But the callsites
of the NegateIfPossible give doubles. This CL fixes it
by static_cast<int32_t> on the doubles.

Bug: 1102403
Change-Id: I3ba3097ce99e5b955a885c84c66471b3dc7dd589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289613
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790496}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eab29100a6aec8a1262b180474109df7f6422e13