Use alpha_type stored with SharedImage for raster CopySubTexture
In the past we had to pass in the alpha type of a SharedImage when
copying it. Recent CL's have added this information to the SharedImage
itself so now we can avoid redundancy and set the alpha type once
when creating the SharedImage. This CL removes
|unpack_premultiply_alpha| from RasterDecoder::CopySubTextureINTERNAL
in favor of using the alpha state stored with the image.

Bug: 1034086
Change-Id: I592bb7f6629212905614fcadf4f8717993923490
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310971
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Nathan Zabriskie <nazabris@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#790597}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5e53f923c462c520584a90a49076cbdf9335ac6