Increment MatchResult::current_tree_order_ by one for each tree scope
This is effectively what we're doing already, since the size of the
author_range_ends_ vector just increased by one (and the size of that
vector never decreases).

Setting the current_tree_order_ when finishing user rules is not
necessary. I don't know why we did this. Possibly it's a remnant of
a previous solution.

Overall, what matters is that current_tree_order_ is monotonically
increasing for each call to FinishAddingAuthorRulesForTreeScope.

This prepares the ua/user/author ranges for removal in a subsequent CL.

Change-Id: I99947e5bfdf727ba772d296ecab65b424625dd91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310495
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791167}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31f328e3e0f04b3d94589c7b1f43d9cfbe979ce3