Change String approach for the identifiability study
Due to the extension of certain templates, the current header declaring
the WTF::String overload of IdentifiabilityDigestHelper must be included
before the main file, which complicates adding new users. We are also
unable to add the template to the main file directly as that would be a
layering violation and as we aim to reduce the usage of complicated
templates.

We thus change the approach to no longer define a template overload and
to instead define a helper that converts the WTF::String into an
IdentifiabilityToken. This helper is moved to blink/platform so that it
is available outside of canvas2d.

The canvas IdentifiabilityStudyHelper and its callers are modified
correspondingly. They now use an IdentifiableTokenBuilder to
incrementally compute its IdentifiableToken.

Bug: 1102935
Change-Id: I5e7a56f629c9a5e15a07ec617932db9a4ef0b90f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212974
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791330}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fd63f34cf0930694548b9e4de015c2f3fff40e3